@@ -6,7 +6,7 @@
     "@emotion/react": "^11.8.2",
     "@emotion/styled": "^11.8.1",
     "@mui/icons-material": "^5.5.1",
-    "@mui/material": "^5.5.3",
+    "@mui/material": "^5.6.0",
     "@mui/styled-engine": "^5.5.2",
     "@mui/styled-engine-sc": "^5.5.2",
     "@testing-library/jest-dom": "^5.16.3",
@@ -18,8 +18,8 @@
     "express": "^4.17.3",
     "graphql": "^16.3.0",
     "mongoose": "^6.2.9",
-    "react": "^17.0.0",
-    "react-dom": "^17.0.0",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-scripts": "5.0.0",
     "styled-components": "^5.3.5",
     "web-vitals": "^2.1.4"
